fix: Passing an int to V8.NewValue
Weird, this wasn't caught by a test. Should add test, but rather get the
fix out now.

+++ scripting/v8host/es_wrapper.go
@@ -148,7 +148,7 @@ func (w converters) toNullableDOMString(ctx *V8ScriptContext, str *string) (*v8.
 }
 
 func (w converters) toUnsignedLong(ctx *V8ScriptContext, val int) (*v8.Value, error) {
-	return v8.NewValue(ctx.host.iso, val)
+	return v8.NewValue(ctx.host.iso, uint32(val))
 }
 
 func (w converters) toAny(ctx *V8ScriptContext, val string) (*v8.Value, error) {
